pyproject.toml
[project]
name = "ableton-mcp-extended"
version = "1.0.0"
description = "Ableton Live integration through the Model Context Protocol with ElevenLabs and extended features"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "uisato"}
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "mcp[cli]>=1.3.0",
    "elevenlabs>=0.2.26",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "aiohttp>=3.9.0",
    "websockets>=12.0",
]

[project.scripts]
ableton-mcp-extended = "MCP_Server.server:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["MCP_Server", "elevenlabs_mcp", "AbletonMCP_UDP", "localai_mcp", "comfyui_mcp", "uvr5_mcp", "rvc_mcp"]

[project.urls]
"Homepage" = "https://github.com/uisato/ableton-mcp-extended"
"Bug Tracker" = "https://github.com/uisato/ableton-mcp-extended/issues"

[project.optional-dependencies]
xy_controller = [
    "pynput>=1.7.6",
    "screeninfo>=0.8.1",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
]